Fix cdata issue on non-strings
Prior to this commit any attempt to build xml with a value that was not
a string would eventually cause an indexOf error when checking if there
was a cdata requirement. Below is an example with the error that it
generates.

The example:

  var xml2js = require('xml2js');
  var obj = {name: "Super", Surname: "Man > Woman", age: 23};
  var builder = new xml2js.Builder({cdata: true});
  var xml = builder.buildObject(obj);
  console.log(xml);

and the error:

  xml2js/lib/xml2js.js:32
    return entry.indexOf('&') >= 0 || entry.indexOf('>') >= 0 ||
    entry.indexOf

The fix is to simply check that it is a string before checking that it
requires cdata.