Fix chaining w/ processors that don't return strings
If using one of the built-in processors w/ a processor that
does not return a string, they'd break because they always
expect a string as input. While it is possible to workaround
this problem by adjusting the order in which processors get
executed, I think this is a nicer solution.