Return a Promise from parseString
Legacy [optional] callback based handling is still supported, a
Promise is now returned for use by the cool kids.

This has some implication, but I don't think this breaks any specs,
just some hacks...

Let me explain.

`parseString` is designed as asynchronous, as such it should
return immediately.

However some tests rely on try/catch block around this asynchronous
function, which is a bit fishy and cannot work anymore when we
return a Promise (which now guarantees that `parseString` returns
immediately).

Userland exceptions should be caught and emited as errors
(with a tag like: `"Exception thrown in callback: " + ...`)

As such:

'test error throwing' was disabled.

Since we return a promise, the error callback cannot be called
synchronously. Callbacks are generally assumed to live in the
asynchronous world.

'test no error event after end'

This test assumes that everything is synchronous when it should not.
It's not failing because it's not testing anything anymore, shoud
we remove the test?