Use $source to restore the original XML nodes order
Quite a refactoring the Builder class.

It is now a 2 steps process:

1. Preprocessing

Convert nodes to a unified format.
In this format, children are extracted as an array, then sorted
by their position in the original XML document (if made available
using the `sourcemap` option).

2. XML building

The code here has been simplified as special cases are handled at
processing stage. It build the XML.