Add reference to isomorphic-xml2js
I created an implementation of about 80-85% of xml2js functionality which bundles to about 5% of the size of xml2js in the browser, and simply exports xml2js when `require`d in a nodejs environment. I'm hoping this can be of value to people, so please consider adding a reference to the project in your README. Thanks for your consideration!